# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cybersecurity Dork Library API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /search:
    get:
      operationId: searchDorks
      tags: [search]
      summary: Full-text search with filters
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
        - name: intentType
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
  /search/suggestions:
    get:
      operationId: getSearchSuggestions
      tags: [search]
      summary: Get search autocomplete suggestions
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /sources:
    get:
      operationId: listSources
      tags: [sources]
      summary: List sources
      responses:
        "200":
          description: Source list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Source"
    post:
      operationId: createSource
      tags: [sources]
      summary: Create a source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSourceBody"
      responses:
        "201":
          description: Created source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
  /sources/{id}:
    put:
      operationId: updateSource
      tags: [sources]
      summary: Update a source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSourceBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
    delete:
      operationId: deleteSource
      tags: [sources]
      summary: Delete a source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /ingestion/run:
    post:
      operationId: runIngestionJob
      tags: [ingestion]
      summary: Trigger an ingestion job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunIngestionBody"
      responses:
        "202":
          description: Job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionJob"
  /ingestion/jobs:
    get:
      operationId: listIngestionJobs
      tags: [ingestion]
      summary: List ingestion jobs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionJobListResponse"
  /raw-content:
    get:
      operationId: listRawContent
      tags: [rawContent]
      summary: List raw content
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: sourceId
          in: query
          schema:
            type: integer
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Raw content list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RawContentListResponse"
  /extract:
    post:
      operationId: extractDorks
      tags: [extractedDorks]
      summary: Extract dork candidates from raw content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractBody"
      responses:
        "200":
          description: Extraction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractResult"
  /extracted-dorks:
    get:
      operationId: listExtractedDorks
      tags: [extractedDorks]
      summary: List extracted dork candidates
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Extracted dork list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractedDorkListResponse"
  /extracted-dorks/{id}/import:
    post:
      operationId: importExtractedDork
      tags: [extractedDorks]
      summary: Import an extracted dork into the main library
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Imported dork
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dork"
  /extracted-dorks/{id}/ignore:
    post:
      operationId: ignoreExtractedDork
      tags: [extractedDorks]
      summary: Mark extracted dork as ignored
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ignored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    Dork:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        queryTemplate:
          type: string
        optimizedQuery:
          type: string
        description:
          type: string
        usageContext:
          type: string
        intentType:
          type: string
        difficulty:
          type: string
        sourceType:
          type: string
        status:
          type: string
        viewsCount:
          type: integer
        copyCount:
          type: integer
        primaryCategoryId:
          type: integer
        primaryCategory:
          $ref: "#/components/schemas/Category"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        operators:
          type: array
          items:
            $ref: "#/components/schemas/Operator"
        platforms:
          type: array
          items:
            $ref: "#/components/schemas/Platform"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, queryTemplate, status, createdAt]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        icon:
          type: string
        color:
          type: string
        depthLevel:
          type: integer
        displayOrder:
          type: integer
        isActive:
          type: boolean
        parentCategoryId:
          type: integer
        dorkCount:
          type: integer
        subcategories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        createdAt:
          type: string
      required: [id, name, slug, depthLevel]
...
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        tagType:
          type: string
        description:
          type: string
        dorkCount:
          type: integer
      required: [id, name, slug]
...
    Operator:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        syntax:
          type: string
        description:
          type: string
        exampleUsage:
          type: string
        dorkCount:
          type: integer
      required: [id, name, syntax]
...
    Platform:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        platformType:
          type: string
        description:
          type: string
        dorkCount:
          type: integer
      required: [id, name, slug]
...
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/Dork"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        query:
          type: string
      required: [results, total, page, limit, totalPages, query]
    Source:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sourceType:
          type: string
        baseUrl:
          type: string
        fetchMethod:
          type: string
        isActive:
          type: boolean
        lastFetchedAt:
          type: string
        jobCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, sourceType, isActive]
    CreateSourceBody:
      type: object
      properties:
        name:
          type: string
        sourceType:
          type: string
        baseUrl:
          type: string
        fetchMethod:
          type: string
        isActive:
          type: boolean
      required: [name, sourceType]
    RunIngestionBody:
      type: object
      properties:
        sourceId:
          type: integer
        jobType:
          type: string
      required: [sourceId]
    IngestionJob:
      type: object
      properties:
        id:
          type: integer
        sourceId:
          type: integer
        source:
          $ref: "#/components/schemas/Source"
        jobType:
          type: string
        status:
          type: string
        startedAt:
          type: string
        finishedAt:
          type: string
        itemsFound:
          type: integer
        itemsProcessed:
          type: integer
        errorMessage:
          type: string
        createdAt:
          type: string
      required: [id, sourceId, status]
    IngestionJobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/IngestionJob"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [jobs, total, page, limit, totalPages]
    RawContent:
      type: object
      properties:
        id:
          type: integer
        sourceId:
          type: integer
        title:
          type: string
        rawText:
          type: string
        sourceUrl:
          type: string
        contentHash:
          type: string
        discoveredAt:
          type: string
        extractedDorkCount:
          type: integer
      required: [id, sourceId, contentHash]
    RawContentListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RawContent"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [items, total, page, limit, totalPages]
    ExtractBody:
      type: object
      properties:
        rawContentId:
          type: integer
        rawText:
          type: string
      required: [rawText]
    ExtractResult:
      type: object
      properties:
        candidates:
          type: array
          items:
            type: string
        count:
          type: integer
      required: [candidates, count]
    ExtractedDork:
      type: object
      properties:
        id:
          type: integer
        rawContentId:
          type: integer
        rawQuery:
          type: string
        normalizedQuery:
          type: string
        optimizedQuery:
          type: string
        extractedTitle:
          type: string
        detectedCategory:
          type: string
        detectedIntent:
          type: string
        confidenceScore:
          type: number
        processingStatus:
          type: string
        createdAt:
          type: string
      required: [id, rawQuery, processingStatus]
    ExtractedDorkListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ExtractedDork"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [items, total, page, limit, totalPages]
...