# lib/api-zod/src/generated/types/extractedDork.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cybersecurity Dork Library API
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractedDork {
  id: number;
  rawContentId?: number;
  rawQuery: string;
  normalizedQuery?: string;
  optimizedQuery?: string;
  extractedTitle?: string;
  detectedCategory?: string;
  detectedIntent?: string;
  confidenceScore?: number;
  processingStatus: string;
  createdAt?: string;
}