# lib/api-zod/src/generated/types/rawContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cybersecurity Dork Library API
 * OpenAPI spec version: 0.1.0
 */

export interface RawContent {
  id: number;
  sourceId: number;
  title?: string;
  rawText?: string;
  sourceUrl?: string;
  contentHash: string;
  discoveredAt?: string;
  extractedDorkCount?: number;
}